Restricted some views so they don't work if the method of the request is not the right one

+++ whatsapp_clone/chat/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, redirect
 from django.contrib.auth.decorators import login_required
 from django.utils import timezone
 from django.db.models import Max
+from django.http import HttpResponseNotAllowed
 from .models import User, Chat, Contact, Message, Status
 from .forms import ChatForm, ContactForm, MessageForm, StatusForm
 from typing import Union
@@ -34,37 +35,45 @@ def chat(request):
 
 # htmx
 def get_chats(request):
-    #chats = request.user.chats.order_by("last_message_date")
-    chats = request.user.chats.all()
-    #chats = request.user.chats.annotate(last_message_date=Max('message__date')).order_by('-last_message_date')
-    user_chats = []
-    for chat in chats:
-        contact = get_contact_in_chat(chat, request.user)
-        if contact:
-            if not contact.archived:
+    if request.method == "GET":
+        # chats = request.user.chats.order_by("last_message_date")
+        chats = request.user.chats.all()
+        # chats = request.user.chats.annotate(last_message_date=Max('message__date')).order_by('-last_message_date')
+        user_chats = []
+        for chat in chats:
+            contact = get_contact_in_chat(chat, request.user)
+            if contact:
+                if not contact.archived:
+                    user_chats.append(chat)
+            # if no contact is found it means is an unknow phone
+            # display it as normal chat
+            else:
                 user_chats.append(chat)
-        # if no contact is found it means is an unknow phone
-        # display it as normal chat
-        else:
-            user_chats.append(chat)
 
-    return render(
-        request, "layouts/partials/components/chats.html", {"chats": user_chats}
-    )
+        return render(
+            request, "layouts/partials/components/chats.html", {"chats": user_chats}
+        )
+    else:
+        return HttpResponseNotAllowed(["GET"])
 
 
 def get_archived_chats(request):
-    chats = request.user.chats.all()
-    user_archived_chats = []
-    for chat in chats:
-        contact = get_contact_in_chat(chat, request.user)
-        if contact:
-            if contact.archived:
-                user_archived_chats.append(chat)
+    if request.method == "GET":
+        chats = request.user.chats.all()
+        user_archived_chats = []
+        for chat in chats:
+            contact = get_contact_in_chat(chat, request.user)
+            if contact:
+                if contact.archived:
+                    user_archived_chats.append(chat)
 
-    return render(
-        request, "layouts/partials/archived_chats.html", {"chats": user_archived_chats}
-    )
+        return render(
+            request,
+            "layouts/partials/archived_chats.html",
+            {"chats": user_archived_chats},
+        )
+    else:
+        return HttpResponseNotAllowed(["GET"])
 
 
 def archive_chat(request, chat_id, archive):
@@ -83,53 +92,70 @@ def archive_chat(request, chat_id, archive):
 
 
 def display_user_ui(request):
-    chats = request.user.chats.all()
-    contacts = request.user.contact_set.all().order_by("name")
+    if request.method == "GET":
+        chats = request.user.chats.all()
+        contacts = request.user.contact_set.all().order_by("name")
 
-    return render(
-        request,
-        "layouts/partials/user_interface.html",
-        {
-            "chats": chats,
-            "contacts": contacts,
-        },
-    )
+        return render(
+            request,
+            "layouts/partials/user_interface.html",
+            {
+                "chats": chats,
+                "contacts": contacts,
+            },
+        )
+    else:
+        return HttpResponseNotAllowed(["GET"])
 
 
 def display_chat(request, pk):
-    chat = Chat.objects.get(id=pk)
-    chat_messages = chat.message_set.all().order_by("date")
-    for message in chat_messages.iterator():
-        if not message.read and message.sender_user.pk != request.user.id:
-            message.read = True
-            message.save()
- 
-    return render(
-        request,
-        "layouts/partials/selected-chat.html",
-        {"chat": chat, "messages": chat_messages},
-    )
+    if request.method == "GET":
+        chat = Chat.objects.get(id=pk)
+        chat_messages = chat.message_set.all().order_by("date")
+        for message in chat_messages.iterator():
+            if not message.read and message.sender_user.pk != request.user.id:
+                message.read = True
+                message.save()
+
+        return render(
+            request,
+            "layouts/partials/selected-chat.html",
+            {"chat": chat, "messages": chat_messages},
+        )
+    else:
+        return HttpResponseNotAllowed(["GET"])
 
 
 def delete_chat(request, pk):
-    chat = Chat.objects.get(id=pk)
-    chat.delete()
-    chats = request.user.chats.all()
-    return render(request, "layouts/partials/components/chats.html", {"chats": chats})
+    if request.method == "DELETE":
+        chat = Chat.objects.get(id=pk)
+        chat.delete()
+        chats = request.user.chats.all()
+        return render(
+            request, "layouts/partials/components/chats.html", {"chats": chats}
+        )
+    else:
+        return HttpResponseNotAllowed(["DELETE"])
 
 
 def get_contacts(request):
-    contacts = request.user.contact_set.all()
-    return render(
-        request, "layouts/partials/components/contacts.html", {"contacts": contacts}
-    )
+    if request.method == "GET":
+        contacts = request.user.contact_set.all()
+        return render(
+            request, "layouts/partials/components/contacts.html", {"contacts": contacts}
+        )
+    else:
+        return HttpResponseNotAllowed(["GET"])
 
 
 def get_contact(request, pk):
-    contact = Contact.objects.get(id=pk)
-    return render(
-        request, "layouts/partials/components/contact.html", {"contact": contact}
-    )
+    if request.method == "GET":
+        contact = Contact.objects.get(id=pk)
+        return render(
+            request, "layouts/partials/components/contact.html", {"contact": contact}
+        )
+    else:
+        return HttpResponseNotAllowed(["GET"])
 
 
 def edit_contact(request, pk):
@@ -144,7 +170,7 @@ def edit_contact(request, pk):
                 "contact_form": contact_form,
             },
         )
-    else:
+    elif request.method == "POST":
         contact = Contact.objects.get(id=pk)
         contact_form = ContactForm(request.POST, instance=contact)
         if contact_form.is_valid():
@@ -152,22 +178,30 @@ def edit_contact(request, pk):
         return render(
             request, "layouts/partials/components/contact.html", {"contact": contact}
         )
+    else:
+        return HttpResponseNotAllowed(["GET", "POST"])
 
 
 def delete_contact(request, pk):
-    contact = Contact.objects.get(id=pk)
-    contact.delete()
-    contacts = request.user.contact_set.all()
-    return render(
-        request, "layouts/partials/components/contacts.html", {"contacts": contacts}
-    )
+    if request.method == "DELETE":
+        contact = Contact.objects.get(id=pk)
+        contact.delete()
+        contacts = request.user.contact_set.all()
+        return render(
+            request, "layouts/partials/components/contacts.html", {"contacts": contacts}
+        )
+    else:
+        return HttpResponseNotAllowed(["DELETE"])
 
 
 def get_message(request, pk):
-    message = Message.objects.get(id=pk)
-    return render(
-        request, "layouts/partials/components/message.html", {"message": message}
-    )
+    if request.method == "GET":
+        message = Message.objects.get(id=pk)
+        return render(
+            request, "layouts/partials/components/message.html", {"message": message}
+        )
+    else:
+        return HttpResponseNotAllowed(["GET"])
 
 
 def edit_message(request, pk):
@@ -182,7 +216,7 @@ def edit_message(request, pk):
                 "message_form": message_form,
             },
         )
-    else:
+    elif request.method == "POST":
         message_instance = Message.objects.get(id=pk)
         message_form = MessageForm(request.POST, instance=message_instance)
         if message_form.is_valid():
@@ -192,45 +226,56 @@ def edit_message(request, pk):
             "layouts/partials/components/message.html",
             {"message": message_instance},
         )
+    else:
+        return HttpResponseNotAllowed(["GET", "POST"])
 
 
 def delete_message(request, chat_id, message_id):
-    message_instance = Message.objects.get(id=message_id)
-    message_instance.delete()
-    chat = Chat.objects.get(id=chat_id)
-    chat_messages = chat.message_set.all().order_by("date")
-    return render(
-        request,
-        "layouts/partials/components/messages.html",
-        {"chat": chat, "messages": chat_messages},
-    )
+    if request.method == "DELETE":
+        message_instance = Message.objects.get(id=message_id)
+        message_instance.delete()
+        chat = Chat.objects.get(id=chat_id)
+        chat_messages = chat.message_set.all().order_by("date")
+        return render(
+            request,
+            "layouts/partials/components/messages.html",
+            {"chat": chat, "messages": chat_messages},
+        )
+    else:
+        return HttpResponseNotAllowed(["DELETE"])
 
 
 def update_chat_form(request):
-    contacts = request.user.contact_set.all()
-    return render(
-        request, "layouts/partials/chat_form_elements.html", {"contacts": contacts}
-    )
+    if request.method == "GET":
+        contacts = request.user.contact_set.all()
+        return render(
+            request, "layouts/partials/chat_form_elements.html", {"contacts": contacts}
+        )
+    else:
+        return HttpResponseNotAllowed(["GET"])
 
 
 def get_statuses(request):
-    contacts_statuses = get_contacts_statuses(request.user, False)
-    muted_contacts_statuses = get_contacts_statuses(request.user, True)
+    if request.method == "GET":
+        contacts_statuses = get_contacts_statuses(request.user, False)
+        muted_contacts_statuses = get_contacts_statuses(request.user, True)
 
-    # Query for statuses uploaded by the user
-    user_statuses = Status.objects.filter(uploaded_by=request.user).order_by(
-        "upload_date"
-    )
+        # Query for statuses uploaded by the user
+        user_statuses = Status.objects.filter(uploaded_by=request.user).order_by(
+            "upload_date"
+        )
 
-    return render(
-        request,
-        "layouts/partials/statuses.html",
-        {
-            "user_statuses": user_statuses,
-            "contacts_with_statuses": contacts_statuses,
-            "muted_contacts_with_statuses": muted_contacts_statuses,
-        },
-    )
+        return render(
+            request,
+            "layouts/partials/statuses.html",
+            {
+                "user_statuses": user_statuses,
+                "contacts_with_statuses": contacts_statuses,
+                "muted_contacts_with_statuses": muted_contacts_statuses,
+            },
+        )
+    else:
+        return HttpResponseNotAllowed(["GET"])
 
 
 def mute_contact_statuses(request, contact_id: Union[str, int], mute: bool):
@@ -247,24 +292,27 @@ def mute_contact_statuses(request, contact_id: Union[str, int], mute: bool):
 
 
 def create_status(request):
-    # gets the text and the image for the status creation
-    text = request.POST.get("text")
-    image = request.FILES.get("image")
-    print(f"TEXT:{text}\nIMAGE:{image}")
-    # new status creation
-    if image or text:
-        new_status = Status.objects.create(
-            uploaded_by=request.user, upload_date=timezone.now()
-        )
-        if text:
-            new_status.text = text
-            new_status.save()
-
-        if image:
-            new_status.image.save(
-                f"user_status_{new_status.id}.{image.content_type.split('/')[-1]}",
-                image,
+    if request.method == "POST":
+        # gets the text and the image for the status creation
+        text = request.POST.get("text")
+        image = request.FILES.get("image")
+        print(f"TEXT:{text}\nIMAGE:{image}")
+        # new status creation
+        if image or text:
+            new_status = Status.objects.create(
+                uploaded_by=request.user, upload_date=timezone.now()
             )
-            new_status.save()
+            if text:
+                new_status.text = text
+                new_status.save()
 
-    return redirect("statuses")
+            if image:
+                new_status.image.save(
+                    f"user_status_{new_status.id}.{image.content_type.split('/')[-1]}",
+                    image,
+                )
+                new_status.save()
+
+        return redirect("statuses")
+    else:
+        return HttpResponseNotAllowed(["POST"])
